requirements.txt
pyrogram>=2.0.30
tgcrypto
pymongo[srv]==3.12.3
motor==2.5.1
marshmallow==3.14.1
umongo==3.0.1
requests
bs4
cinemagoer
datetime
pytz
sticker
Pyromod
youtube-dl
youtube_search
requests
googletrans==3.1.0a0
telegraph
beautifulsoup4
aiohttp==3.7.4
gTTS
googletrans==3.1.0a0
NumPy
glitch_this
opencv-python-headless
youtube-search-python==1.4.6
yt-dlp
pypng
Pillow
wget
aiofiles
psutil
wheel
humanize
ytthumb
pypdf2
PyPDF4
pyshorteners
ffmpeg-python
TgCrypto
py-tgcalls
python-dotenv
youtube_search_python
requests
aiohttp
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg-python
ffmpeg
Pillow
ujson
